Removed email from user settings. Upvote update sort of works...not really
